+++ src/common/mutable.rs
@@ -10,7 +10,7 @@ use crate::Id;
 use super::PutMutableRequestArguments;
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-/// [Bep_0044](https://www.bittorrent.org/beps/bep_0044.html)'s Mutable item.
+/// [BEP_0044](https://www.bittorrent.org/beps/bep_0044.html)'s Mutable item.
 pub struct MutableItem {
     /// hash of the key and optional salt
     target: Id,